Fix the budget status text to update dynamically in the GUI class
Create a new functions update_lodging_cost and update_incidental_cost
inside the GUI class to update these values in the Budget object when
the slider is moved.

Refactorr the budgetStatus label to be set by udpate_budgetStatus function,
which gets the latest months_remaining and burndown from the Budget object
and update the label string in the UI.

update_lodging_costs and update_incidental_cost call update_budgetStatus.

Adds a text file for notes for my blog post on the builder and factory
design patterns

TODO: Refactor __init__ and createWidgets to pull widget building out of
single function.