Add class for recurring expenses and GUI tool to generate new sliders from recurring expense file
Add class Expense_Rec which takes name, from_, to, and current values for
a given expense.  This is the first phase of the refactor away from explicitely
enumerating the slider names and values.

Add method create_newExpenseRec in the GUI class to which generates a slider
from an Expense_Rec object. Slider does not currently change any data on
the back-end.